requirements.txt
autopep8==1.4.4
certifi==2019.3.9
crispy-bootstrap4==2022.1
chardet==3.0.4
defusedxml==0.6.0
Django==4.2.1
django-allauth==0.54.0
django-cors-headers==4.0.0
django-countries==7.5.1
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
idna==2.8
oauthlib==3.0.1
pep8==1.7.1
Pillow==9.5.0
pycodestyle==2.5.0
pyodbc==4.0.39
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
sqlparse==0.4.4
stripe==2.27.0
urllib3==1.24.2
